Move the bread crumb navigation to its own component
